# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "The one JSON error shape returned by every API error boundary.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fieldErrors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Fielderrors",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "fieldErrors",
          "requestId",
          "retryable"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "PreferencesResponse": {
        "additionalProperties": false,
        "properties": {
          "language": {
            "title": "Language",
            "type": "string"
          },
          "theme": {
            "title": "Theme",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "theme",
          "language",
          "version"
        ],
        "title": "PreferencesResponse",
        "type": "object"
      },
...
      "UpdatePreferencesRequest": {
        "additionalProperties": false,
        "properties": {
          "language": {
            "anyOf": [
              {
                "enum": [
                  "zh-CN",
                  "en"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "theme": {
            "anyOf": [
              {
                "enum": [
                  "light",
                  "dark",
                  "system"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "UpdatePreferencesRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Secure, HttpOnly current-account session cookie.",
        "in": "cookie",
        "name": "__Host-qg_session",
        "type": "apiKey"
      },
      "SessionCsrf": {
        "description": "Readable session-bound CSRF cookie value echoed by the same-origin client.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "QuantGym API",
    "version": "2.0.0-preview.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/preferences": {
      "patch": {
        "operationId": "updatePreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request proof or permission denied"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Version or idempotency conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": [],
            "SessionCsrf": []
          }
        ],
        "summary": "Update one current-account preference with version checking",
        "tags": [
          "preferences"
        ]
      }
    },
...
  }
}